pyproject.toml
[project]
name = "mcp-todo-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "mcp[cli]>=1.0.0",
    "uvicorn>=0.32.1",
    "openai>=1.66.3",
    "streamlit>=1.43.2",
    "colorama>=0.4.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp-todo-app"]
